feat: move to MEGASTROKE comparison
Use an external reference dataset for the real-data evaluation of
MaxGCP. This avoids the issues we were having with full-size ICD codes
in the UK Biobank as our reference (i.e. that the "target" phenotype is
still defined using a noisy definition), and it allows us to compare out
method, even when using the full cohort.